Widget part of calendar library is moved into calendar package
