+++ matrix-meetings-widget/src/components/meetings/MeetingCardShareMeetingContent/useMeetingEmail.ts
@@ -14,10 +14,12 @@
  * limitations under the License.
  */
 
-import { parseICalDate } from '@nordeck/matrix-meetings-calendar';
+import {
+  isRecurringCalendarSourceEntry,
+  parseICalDate,
+} from '@nordeck/matrix-meetings-calendar';
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import { isRecurringCalendarSourceEntry } from '../../../lib/utils';
 import { Meeting } from '../../../reducer/meetingsApi';
 import { fullNumericDateFormat } from '../../common/DateTimePickers';
 import { formatRRuleText } from '../RecurrenceEditor/utils';
