+++ matrix-meetings-widget/src/components/meetings/MeetingDetails/MeetingDetailsContent/MeetingDetailsContent.tsx
@@ -18,8 +18,8 @@ import EventRepeatIcon from '@mui/icons-material/EventRepeat';
 import LinkIcon from '@mui/icons-material/Link';
 import TodayIcon from '@mui/icons-material/Today';
 import { Box, DialogContent, Grid, Link, Typography } from '@mui/material';
+import { isRecurringCalendarSourceEntry } from '@nordeck/matrix-meetings-calendar';
 import { useTranslation } from 'react-i18next';
-import { isRecurringCalendarSourceEntry } from '../../../../lib/utils';
 import { Meeting } from '../../../../reducer/meetingsApi';
 import { CopyableTextButton } from '../../../common/CopyableTextButton';
 import { fullLongDateFormat } from '../../../common/DateTimePickers';
