+++ matrix-meetings-widget/src/components/meetings/ScheduleMeetingModal/ScheduleMeetingModal.tsx
@@ -15,9 +15,12 @@
  */
 
 import { useWidgetApi } from '@matrix-widget-toolkit/react';
-import { parseICalDate, toISOString } from '@nordeck/matrix-meetings-calendar';
+import {
+  isRecurringCalendarSourceEntry,
+  parseICalDate,
+  toISOString,
+} from '@nordeck/matrix-meetings-calendar';
 import { useEffect, useMemo, useState } from 'react';
-import { isRecurringCalendarSourceEntry } from '../../../lib/utils';
 import { Meeting } from '../../../reducer/meetingsApi';
 import { EditRecurringMessage } from './EditRecurringMessage';
 import { ScheduleMeeting } from './ScheduleMeeting';
