+++ matrix-meetings-widget/src/components/meetings/ScheduleMeetingModal/editMeetingThunk.tsx
@@ -17,13 +17,13 @@
 import {
   CalendarEntry,
   formatICalDate,
+  normalizeCalendarEntry,
+  overrideCalendarEntries,
 } from '@nordeck/matrix-meetings-calendar';
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import { t } from 'i18next';
 import { DateTime } from 'luxon';
 import { ModalButtonKind } from 'matrix-widget-api';
-import { normalizeCalendarEntry } from '../../../lib/utils';
-import { overrideCalendarEntries } from '../../../lib/utils/calendarUtils/overrideCalendarEntries';
 import { meetingBotApi } from '../../../reducer/meetingBotApi';
 import {
   Meeting,
