--- matrix-meetings-widget/src/lib/utils/calendarUtils/helpers.ts
@@ -1,52 +1,0 @@
-/*
- * Copyright 2022 Nordeck IT + Consulting GmbH
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { CalendarEntry } from '@nordeck/matrix-meetings-calendar';
-import { DateTime } from 'luxon';
-import { CalendarSourceEntries } from './calculateCalendarEvents';
-
-export function isSingleCalendarSourceEntry(
-  entries?: CalendarSourceEntries,
-): entries is [Omit<CalendarEntry, 'rrule' | 'exdate' | 'recurrenceId'>] {
-  return entries?.length === 1 && entries[0].rrule === undefined;
-}
-
-export function isRecurringCalendarSourceEntry(
-  entries?: CalendarSourceEntries,
-): entries is
-  | [CalendarEntry & { rrule: string }]
-  | [
-      CalendarEntry & { rrule: string },
-      Required<Omit<CalendarEntry, 'rrule' | 'exdate'>>,
-    ] {
-  return (
-    typeof entries?.[0].rrule === 'string' &&
-    (entries[1] === undefined || entries[1].recurrenceId !== undefined)
-  );
-}
-
-export function createTimeFilter(fromDate: string, toDate?: string) {
-  const fromDateTime = DateTime.fromISO(fromDate);
-  const toDateTime =
-    toDate !== undefined ? DateTime.fromISO(toDate) : undefined;
-
-  return ({ startTime, endTime }: { startTime: string; endTime: string }) => {
-    return (
-      fromDateTime < DateTime.fromISO(endTime) &&
-      (!toDateTime || DateTime.fromISO(startTime) <= toDateTime)
-    );
-  };
-}
