--- matrix-meetings-widget/src/lib/utils/calendarUtils/isRecurringMeeting.ts
@@ -1,21 +1,0 @@
-/*
- * Copyright 2022 Nordeck IT + Consulting GmbH
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { CalendarEntry } from '@nordeck/matrix-meetings-calendar';
-
-export function isRecurringMeeting(calendar: CalendarEntry[]): boolean {
-  return calendar.length > 1 || calendar.some((e) => e.rrule);
-}
