+++ packages/calendar/src/utils/calendarUtils/helpers.ts
@@ -14,8 +14,10 @@
  * limitations under the License.
  */
 
+import { DateTime } from 'luxon';
 import { Options } from 'rrule';
 import { CalendarEntry, DateTimeEntry } from '../../model';
+import { CalendarSourceEntries } from './calculateCalendarEvents';
 
 export function isSingleEntry(
   entry: CalendarEntry,
@@ -39,4 +41,37 @@ export function isRRuleOverrideEntry(
 
 export function isFiniteSeries(rruleOptions: Partial<Options>): boolean {
   return rruleOptions.count !== undefined || rruleOptions.until !== undefined;
+}
+
+export function isSingleCalendarSourceEntry(
+  entries?: CalendarSourceEntries,
+): entries is [Omit<CalendarEntry, 'rrule' | 'exdate' | 'recurrenceId'>] {
+  return entries?.length === 1 && entries[0].rrule === undefined;
+}
+
+export function isRecurringCalendarSourceEntry(
+  entries?: CalendarSourceEntries,
+): entries is
+  | [CalendarEntry & { rrule: string }]
+  | [
+      CalendarEntry & { rrule: string },
+      Required<Omit<CalendarEntry, 'rrule' | 'exdate'>>,
+    ] {
+  return (
+    typeof entries?.[0].rrule === 'string' &&
+    (entries[1] === undefined || entries[1].recurrenceId !== undefined)
+  );
+}
+
+export function createTimeFilter(fromDate: string, toDate?: string) {
+  const fromDateTime = DateTime.fromISO(fromDate);
+  const toDateTime =
+    toDate !== undefined ? DateTime.fromISO(toDate) : undefined;
+
+  return ({ startTime, endTime }: { startTime: string; endTime: string }) => {
+    return (
+      fromDateTime < DateTime.fromISO(endTime) &&
+      (!toDateTime || DateTime.fromISO(startTime) <= toDateTime)
+    );
+  };
 }
