+++ packages/calendar/src/utils/calendarUtils/index.ts
@@ -14,11 +14,20 @@
  * limitations under the License.
  */
 
+export { calculateCalendarEvents } from './calculateCalendarEvents';
+export type { CalendarSourceEntries } from './calculateCalendarEvents';
+export { deleteCalendarEvent } from './deleteCalendarEvent';
 export { generateRruleSet } from './generateRruleSet';
 export { getCalendarEnd } from './getCalendarEnd';
+export { getCalendarEvent } from './getCalendarEvent';
 export {
   isFiniteSeries,
   isRRuleEntry,
   isRRuleOverrideEntry,
+  isRecurringCalendarSourceEntry,
+  isSingleCalendarSourceEntry,
   isSingleEntry,
 } from './helpers';
+export { isRecurringMeeting } from './isRecurringMeeting';
+export { normalizeCalendarEntry } from './normalizeCalendarEntry';
+export { overrideCalendarEntries } from './overrideCalendarEntries';
