+++ packages/calendar/src/utils/calendarUtils/normalizeCalendarEntry.test.ts
