+++ packages/calendar/src/utils/calendarUtils/normalizeCalendarEntry.ts
@@ -14,12 +14,9 @@
  * limitations under the License.
  */
 
-import {
-  CalendarEntry,
-  formatICalDate,
-  parseICalDate,
-} from '@nordeck/matrix-meetings-calendar';
 import { DateTime } from 'luxon';
+import { CalendarEntry } from '../../model';
+import { formatICalDate, parseICalDate } from '../dateTimeUtils';
 import { calculateCalendarEvents } from './calculateCalendarEvents';
 
 export function normalizeCalendarEntry(
